test/integration: make Version object return mocks
