+++ internal/testutil/server.go
@@ -1,7 +1,6 @@
 package testutil
 
 import (
-	"embed"
 	"fmt"
 	"io/fs"
 	"net/http"
@@ -9,37 +8,6 @@ import (
 	"path/filepath"
 	"strings"
 )
-
-type Version string
-
-// List of all the versions we have testdata for.
-// When adding a new version:
-// - REMEMBER TO ADD IT TO AllVersions() BELOW.
-// - UPDATE LatestVersion() BELOW IF NEEDED
-
-const (
-	Testdata116 = "1_16"
-	Testdata118 = "1_18"
-)
-
-// LatestVersion returns the latest version we have test data for.
-func LatestVersion() Version {
-	return Testdata118
-}
-
-// AllVersions returns a list of versions we have test data for.
-func AllVersions() []Version {
-	return []Version{
-		Testdata116,
-		Testdata118,
-	}
-}
-
-//go:embed data
-var testDataDir embed.FS
-
-// Name of the root folder in embed.FS
-const testDataRootDir = "data"
 
 type Server struct {
 	*httptest.Server
@@ -56,7 +24,7 @@ func (s *Server) ControlPlaneEndpoint(component string) string {
 	return s.Server.URL + "/controlplane/" + strings.ReplaceAll(component, "/", "")
 }
 
-func NewServer(version Version) (*Server, error) {
+func newServer(version Version) (*Server, error) {
 	subversion, err := fs.Sub(testDataDir, filepath.Join(testDataRootDir, string(version)))
 	if err != nil {
 		return nil, fmt.Errorf("opening dir for version %s: %w", version, err)
